added button for hiding bounding box
